+++ CHANGES.md
@@ -15,6 +15,7 @@
 - [ADD] CI の go を 1.13 に上げる
 - [UPDATE] @kdxu 推奨の go version を 1.13 にあげる
 - [UPDATE] @kdxu 先に入室しているユーザーがいる場合 isExistUser をaccept時にtrue にして返す
+- [FIX] @kdxu roomId か clientId が空文字列の場合は reject するように修正する
 
 ## 19.08.0